moves application controller method to concern
